backend/requirements.txt
Flask
sqlalchemy
mysql-connector-python
passlib[bcrypt]
pyjwt